Change uint256S to Int

+++ src/consensus/params.h
@@ -78,7 +78,7 @@ struct Params {
     int64_t nPowMaxAdjustUp;
     int64_t nPowMaxAdjustDown;
     int64_t nUpdateDiffAlgoHeight;
-    uint256 nMinimumChainWork;
+    int nMinimumChainWork;
     uint256 defaultAssumeValid;
 
     int64_t AveragingWindowTimespan() const { return nPowAveragingWindow * nPowTargetSpacing; }
